+++ Semana_3/gobarber/src/hooks/AuthContext.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useCallback, useState } from "react"; // contexto para tornar acessivel entre componentes que nao tenham relacao
+import React, { createContext, useCallback, useState, useContext } from "react"; // contexto para tornar acessivel entre componentes que nao tenham relacao
 import api from "../services/api";
 
 interface SignInCredentias {
@@ -15,11 +15,10 @@ interface AuthContextData {
   user: object;
   token?: string;
   signIn(credentials: SignInCredentias): Promise<void>;
+  signOut(): void;
 }
 
-export const AuthContext = createContext<AuthContextData>(
-  {} as AuthContextData
-); // Para iniciar o contexto sem valor
+const AuthContext = createContext<AuthContextData>({} as AuthContextData); // Para iniciar o contexto sem valor
 
 const AuthProvider: React.FC = ({ children }) => {
   let [auth, setAuth] = useState<AuthState>(() => {
@@ -43,13 +42,30 @@ const AuthProvider: React.FC = ({ children }) => {
 
     setAuth({ token, user });
   }, []);
+
+  const signOut = useCallback(() => {
+    localStorage.removeItem("@GoBarber:token");
+    localStorage.removeItem("@GoBarber:user");
+
+    setAuth({} as AuthState);
+  }, []);
   return (
     <AuthContext.Provider
-      value={{ token: auth.token, user: auth.user, signIn }}
+      value={{ token: auth.token, user: auth.user, signIn, signOut }}
     >
       {children}
     </AuthContext.Provider>
   );
 };
+
+export function useAuth(): AuthContextData {
+  const context = useContext(AuthContext);
+
+  if (!context) {
+    throw new Error(`useAuth must be used within a AuthProvider`);
+  }
+
+  return context;
+}
 
 export default AuthProvider;